docker/whisper/requirements.txt
openai-whisper==20231117
boto3==1.34.162
mysql-connector-python==9.0.0
requests==2.32.3
pymysql==1.1.1
flask==2.3.3
gunicorn==21.2.0